pyproject.toml
[project]
name = "next-message-prediction-public"
version = "0.1.0"
description = "Next message prediction with LLMs"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "anthropic>=0.47.0",
    "google-generativeai>=0.8.4",
    "ipython>=8.32.0",
    "jupyter>=1.1.1",
    "langfuse>=2.59.4",
    "matplotlib>=3.10.0",
    "numpy>=2.2.3",
    "openai>=1.64.0",
    "pandas>=2.2.3",
    "python-dotenv>=1.0.1",
    "seaborn>=0.13.2",
]
